requirements.txt
python-nest==4.0.3
boto3